ref(installer): move locale, mount-option and AUR helpers next to what they use
* uncomment_locale sits with locale_entry_re in localization/utils
* harden_boot_options sits with BootMountOption and has_separate_boot in models/device,
  the test imports it from there instead of reaching into Installer
* pm/aur.py owns the AUR bootstrap package set and run_grimoire_installation,
  schema_gen and the scripts follow
* run_custom_user_commands goes through arch_chroot() instead of the private prefix
